Remove the need to call modifier on exported modifier
It was previously necessary to call Modifier.modifier on the exported
modifier class.  We can remove this requirement based on the input of
@chancancode in #7.  This commit implements the behaviour as we
understood it from this issue.

The deprecation message indicates version 1.0.0 will drop support for
calling Modifier.modifier, this has not been discussed and is subject
to change.

The examples of the README have been updated to not show the
deprecated Modifier.modifier call.

The tests have been duplicated to contain both the calls with
Modifier.modifier as well as those without.  The tests have been
updated mechanically with Emacs Macro's rather than being smart about
them to keep the tests obvious.