Squashed commits from ember-class-based-modifier
This is a result of squashing everything after the v0.5.0 tag up to and
including the v0.9.0 tag, rebased on top of the current master, which is
the v0.5.2 tag.

The original commits are at the [archive/ember-class-based-modifier-fork
branch](https://github.com/sukima/ember-oo-modifiers/tree/archive/ember-class-based-modifier-fork)
as a permanent archive, but the commits themselves are not very
interesting. It's mainly there so that the `v0.8.0` and `v0.9.0` tags
have something to point to.